+++ src/main/java/com/roberto/transactions/OutputReturn.java
@@ -1,13 +1,14 @@
 package com.roberto.transactions;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class OutputReturn {
     private boolean approved;
-    private double newlimit;
+    private BigDecimal newlimit;
     private List<String> deniedReasons;
 
-    public OutputReturn(boolean approved, double newlimit, List<String> deniedReasons) {
+    public OutputReturn(boolean approved, BigDecimal newlimit, List<String> deniedReasons) {
         this.approved = approved;
         this.newlimit = newlimit;
         this.deniedReasons = deniedReasons;
@@ -17,7 +18,7 @@ public class OutputReturn {
         this.approved = approved;
     }
 
-    public void setNewlimit(double newlimit) {
+    public void setNewlimit(BigDecimal newlimit) {
         this.newlimit = newlimit;
     }
 }
